dockerized-sbert/requirements.txt
SQLAlchemy==1.2.14
hdbscan==0.8.26
umap
pandas==1.1.0
fastapi
pymysql==0.10.1
sentence-transformers==0.3.9
psycopg2-binary==2.8.6
boto3
python-box 5.2.0
python-dotenv